ai-ml/requirements.txt
# Core LangChain
langchain>=0.1.0
langchain-huggingface>=0.0.6
langsmith>=0.0.77

# LLM and Embeddings
huggingface_hub>=0.20.0
transformers>=4.36.0
sentence-transformers>=2.2.0
torch>=2.1.0
tiktoken>=0.5.2

# Vector Storage
faiss-cpu>=1.7.4

# Graph Database
neo4j>=5.14.0

# NLP and ML
spacy>=3.7.0
scikit-learn>=1.3.0
nltk>=3.8.0
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.1.0

# API and Web
Flask>=3.0.0
Werkzeug>=3.0.0
flask-restx>=1.3.0
flask-cors>=4.0.0
gunicorn==20.1.0
gevent>=23.9.0

# PDF Processing
PyPDF2>=3.0.0

# Utilities
PyYAML>=6.0.0
requests>=2.31.0
tqdm>=4.66.0
python-dotenv>=1.0.0
typing_extensions>=4.8.0
urllib3>=2.1.0

# Security
pyopenssl>=23.3.0
markupsafe>=2.1.1
psutil